Attempt to detect languages and scripts for releases.
Make the Bandcamp and Qobuz sources call a new Autofill
method on seed.Release to use the Google Cloud Translation
API to try to detect the languages and scripts that are used
in the release and track titles.

If GCP fails, the 'nogcp' build tag is supplied, or unit
tests are being run, a fallback local implementation that
only makes a weak attempt to guess the script is used
instead.

Fixes #5.