Support editing existing releases from URLs.
Add an "existing MBID" input field in yambsd that can be
used to edit an existing release instead of creating a new
one when seeding from a URL. This is intended to make it
easier to add data to Bandcamp albums that were entered in a
prerelease state (e.g. with incomplete or incorrect
tracklists).

I'm also updating online.Fetch() to generate an additional
edit linking directly to /release/<mbid>/add-cover-art when
an MBID has been supplied, to save a few clicks when just
adding a cover image to an existing release.