sources/online/tidal: Choose best country for -country XW.
When querying all countries, fetch tracklists first and use
the country with the most tracks when fetching album info
and credits. The code previously defaulted to US for album
info and credits, but the album may be completely
unavailable there. Fixes #25.

Also add a tiny fetch_data.sh script to make it easier to
add test data for new albums and countries.