sources/online/bandcamp: Skip null release_date properties.
https://tauk.bandcamp.com/track/daydreams-ft-kanika-moore
has a null 'release_date' property, which I hadn't seen
before and which made unmarshaling of the JSON object fail.

Update the jsonDate struct to just use the zero value when
unmarshaling an empty string. At least in the case of this
album, 'publish_date' is still supplied, so we use that
instead.